se corrigen problemas del APIRest, se avanza con la UI al 85%, el area de notas aun no muestra los datos por pantalla, pero si los trae por consola.

+++ app/components/Task.js
@@ -1,20 +1,22 @@
 import React from "react";
 import '/app/app.css';
-import {CSS} from '@dnd-kit/utilities';
 import del from '../images/del.png';
+import editar from '../images/editar.png';
 
 function Task({task, id}) {
 
     const delTas = (id) => {
-        console.log(id)
-        fetch(`http://127.0.0.1:3001/tareas/${id}` , {
-            method: "DELETE",
-            headers: { "Content-Type": "application/json", "Accept": "application/json" },
+        if(confirm("Realmente deseas eliminarla")){
+                fetch(`http://127.0.0.1:3001/api/task/${id}` , {
+                method: "DELETE",
+                headers: { "Content-Type": "application/json", "Accept": "application/json" },
           })
-          .then(resp => resp.json())
-          .then(data => console.log(data))
-          .catch(error => console.error(error));
-       
+            .then(resp => resp.json())
+            .then(data => console.log(data))
+            .catch(error => console.error(error));
+        }
+        
+      
       }
     return(
         <div  id="taskUI">
@@ -22,6 +24,7 @@ function Task({task, id}) {
                 <li>{task.priority}</li>
                 <li>{task.description}</li>
                 <li><button className="btn-del" onClick={()=> delTas(task._id)}><img id="img-task" src={del}/></button> </li>
+                <li><button className="btn-edit"><img id="img-task" src={editar}/></button></li>
                 </ul>
         </div>
     )
